+++ StepsRiverNetwork.py
@@ -13,6 +13,7 @@ class StepsRiverNetwork(base.Component):
     """The component encapsulating the Steps environmental fate module."""
     # RELEASES
     VERSION = base.VersionCollection(
+        base.VersionInfo("2.1.6", "2023-09-13"),
         base.VersionInfo("2.1.5", "2023-09-12"),
         base.VersionInfo("2.1.4", "2023-09-11"),
         base.VersionInfo("2.1.3", "2022-03-03"),
@@ -120,6 +121,8 @@ class StepsRiverNetwork(base.Component):
     VERSION.added("2.1.5", "Creation of repository info during documentation")
     VERSION.added("2.1.5", "Repository info, changelog, contributing note and readme to module")
     VERSION.added("2.1.5", "Repository info to Python runtime environment")
+    VERSION.added("2.1.6", "Scales attribute to global inputs")
+    VERSION.fixed("2.1.6", "Spatial scale of outputs")
 
     def __init__(self, name, observer, store):
         """
@@ -218,64 +221,64 @@ class StepsRiverNetwork(base.Component):
             ),
             base.Input(
                 "MolarMass",
-                (attrib.Class(float, 1), attrib.Unit("g/mol", 1)),
+                (attrib.Class(float, 1), attrib.Unit("g/mol", 1), attrib.Scales("global")),
                 self.default_observer,
                 description="The molar mass of the substance depositing at the water body surface."
             ),
             base.Input(
                 "DT50sw",
-                (attrib.Class(float, 1), attrib.Unit("d", 1)),
+                (attrib.Class(float, 1), attrib.Unit("d", 1), attrib.Scales("global")),
                 self.default_observer,
                 description="""The half-life transformation time in water of the substance depositing at the water body 
                 surface."""
             ),
             base.Input(
                 "DT50sed",
-                (attrib.Class(float, 1), attrib.Unit("d", 1)),
+                (attrib.Class(float, 1), attrib.Unit("d", 1), attrib.Scales("global")),
                 self.default_observer,
                 description="""The half-life transformation time in sediment of the substance depositing at the water 
                 body surface."""
             ),
             base.Input(
                 "KOC",
-                (attrib.Class(float, 1), attrib.Unit("l/kg", 1)),
+                (attrib.Class(float, 1), attrib.Unit("l/kg", 1), attrib.Scales("global")),
                 self.default_observer,
                 description="""The coefficient for equilibrium adsorption in sediment of the substance depositing at 
                 the water body surface."""
             ),
             base.Input(
                 "Temp0",
-                (attrib.Class(float, 1), attrib.Unit("°C", 1)),
+                (attrib.Class(float, 1), attrib.Unit("°C", 1), attrib.Scales("global")),
                 self.default_observer,
                 description="The reference temperature to which the physical and chemical substance values apply."
             ),
             base.Input(
                 "Q10",
-                (attrib.Class(float, 1), attrib.Unit("1", 1)),
+                (attrib.Class(float, 1), attrib.Unit("1", 1), attrib.Scales("global")),
                 self.default_observer,
                 description="The temperature coefficient for chemical reactions of the deposited substance."
             ),
             base.Input(
                 "PlantUptake",
-                (attrib.Class(float, 1), attrib.Unit("1", 1)),
+                (attrib.Class(float, 1), attrib.Unit("1", 1), attrib.Scales("global")),
                 self.default_observer,
                 description="The fraction of pesticide that is taken up by plants."
             ),
             base.Input(
                 "QFac",
-                (attrib.Class(float, 1), attrib.Unit("1", 1)),
+                (attrib.Class(float, 1), attrib.Unit("1", 1), attrib.Scales("global")),
                 self.default_observer,
                 description="The QFac parameter is not documented in the module documentation."
             ),
             base.Input(
                 "ThresholdSW",
-                (attrib.Class(float, 1), attrib.Unit("mg/m³", 1)),
+                (attrib.Class(float, 1), attrib.Unit("mg/m³", 1), attrib.Scales("global")),
                 self.default_observer,
                 description="The minimum surface water concentration that is reported."
             ),
             base.Input(
                 "ThresholdSediment",
-                (attrib.Class(float, 1), attrib.Unit("mg/kg", 1)),
+                (attrib.Class(float, 1), attrib.Unit("mg/kg", 1), attrib.Scales("global")),
                 self.default_observer,
                 description="The minimum sediment concentration that is reported."
             ),
@@ -362,7 +365,7 @@ class StepsRiverNetwork(base.Component):
                 "PEC_SW",
                 store,
                 self,
-                {"data_type": np.float, "scales": "time/hour, space/base_geometry"},
+                {"data_type": np.float, "scales": "time/hour, space/reach"},
                 "The modelled concentration in the water phase.",
                 {
                     "type": np.ndarray,
@@ -375,7 +378,7 @@ class StepsRiverNetwork(base.Component):
                 "MASS_SW",
                 store,
                 self,
-                {"data_type": np.float, "scales": "time/hour, space/base_geometry"},
+                {"data_type": np.float, "scales": "time/hour, space/reach"},
                 "The modelled substance mass in the water phase.",
                 {
                     "type": np.ndarray,
@@ -388,7 +391,7 @@ class StepsRiverNetwork(base.Component):
                 "MASS_SED",
                 store,
                 self,
-                {"data_type": np.float, "scales": "time/hour, space/base_geometry"},
+                {"data_type": np.float, "scales": "time/hour, space/reach"},
                 "The modelled substance mass in sediment.",
                 {
                     "type": np.ndarray,
@@ -401,7 +404,7 @@ class StepsRiverNetwork(base.Component):
                 "MASS_SED_DEEP",
                 store,
                 self,
-                {"data_type": np.float, "scales": "time/hour, space/base_geometry"},
+                {"data_type": np.float, "scales": "time/hour, space/reach"},
                 "The modelled substance mass in deep sediment.",
                 {
                     "type": np.ndarray,
@@ -414,7 +417,7 @@ class StepsRiverNetwork(base.Component):
                 "PEC_SED",
                 store,
                 self,
-                {"data_type": np.float, "scales": "time/hour, space/base_geometry"},
+                {"data_type": np.float, "scales": "time/hour, space/reach"},
                 "The modelled concentration in sediment.",
                 {
                     "type": np.ndarray,